pyserver/functions/zhaofengli-similar_star_face_finder-dev/requirements.txt
absl-py==0.1.13
alembic==0.9.9
astor==0.6.2
backcall==0.1.0
beautifulsoup4==4.5.3
bleach==1.5.0
bokeh==0.12.15
certifi==2018.1.18
chardet==3.0.4
cloudpickle==0.2.2
cycler==0.10.0
Cython==0.25.2
dask==0.17.2
decorator==4.3.0
dill==0.2.7.1
entrypoints==0.2.3
gast==0.2.0
grpcio==1.11.0
h5py==2.6.0
html5lib==0.9999999
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterhub==0.8.1
jupyterlab==0.31.12
jupyterlab-launcher==0.10.5
Keras==2.1.5
llvmlite==0.22.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.8.3
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numba==0.31.0
numexpr==2.6.4
numpy==1.14.2
opencv-python==3.4.0.12
packaging==17.1
pamela==0.3.0
pandas==0.19.2
pandocfilters==1.4.2
parso==0.2.0
patsy==0.4.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
python-editor==1.0.3
python-oauth2==1.1.0
pytz==2018.4
PyYAML==3.12
pyzmq==17.0.0
requests==2.18.4
scikit-image==0.12.3
scikit-learn==0.18.2
scipy==0.19.1
seaborn==0.7.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.18
statsmodels==0.8.0
sympy==1.0
tensorboard==1.6.0
tensorflow==1.6.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
vincent==0.4.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
xlrd==1.1.0
